Added some functions

+++ internal/products/model/productModel.go
@@ -1,10 +1,10 @@
 package model
 
 type Product struct {
-	ID          int    `json:"id" db:"id"`
-	Name        string `json:"name" db:"name"`
-	Description string `json:"description" db:"description"`
-	Image       string `json:"image" db:"image"`
-	Price       int    `json:"price" db:"price"`
-	CategoryID  int    `json:"category_id" db:"category_id"`
+	ID          int     `json:"id" db:"id"`
+	Name        string  `json:"name" db:"name"`
+	Description string  `json:"description" db:"description"`
+	Image       string  `json:"image" db:"image"`
+	Price       float64 `json:"price" db:"price"`
+	CategoryID  int     `json:"category_id" db:"category_id"`
 }
